CLASS DEFINITION: Move test method into `ClassDefinitionTest::class`
The testRedundantUaSortInClassDefinition method that was earlier
defined in a class called `ClaimTest::class`
